move mbtiles to the core
