+++ martin-core/README.md
@@ -11,7 +11,7 @@ The library providing the basic building blocks for [Martin](https://maplibre.or
 # Features
 
 - tiles
-  - [ ] mbtiles
+  - [x] mbtiles
   - [ ] pmtiles
   - [ ] cog
   - [ ] postgres tables
