+++ martin-core/src/tiles/mbtiles/error.rs
@@ -1,5 +1,8 @@
 //! Error types for `MBTiles` operations.
 
+use std::path::PathBuf;
+
+/// Errors that can occur during mbtiles processing operations.
 #[derive(thiserror::Error, Debug)]
 pub enum MbtilesError {
     /// Failed to acquire database connection to `MBTiles` file.
@@ -9,4 +12,12 @@ pub enum MbtilesError {
     /// Wrapper for underlying mbtiles library errors.
     #[error(transparent)]
     MbtilesLibraryError(#[from] mbtiles::MbtError),
+
+    /// IO error.
+    #[error("IO error {0}: {1}")]
+    IoError(std::io::Error, PathBuf),
+
+    /// Unable to parse metadata in file.
+    #[error(r"Unable to parse metadata in file {1}: {0}")]
+    InvalidMetadata(String, PathBuf),
 }
