+++ martin/Cargo.toml
@@ -95,7 +95,7 @@ default = [
 cog = ["dep:png", "dep:tiff"]
 fonts = ["martin-core/fonts"]
 lambda = ["dep:lambda-web"]
-mbtiles = ["dep:mbtiles"]
+mbtiles = ["martin-core/mbtiles"]
 metrics = ["dep:actix-web-prom"]
 pmtiles = ["dep:aws-config", "dep:pmtiles"]
 postgres = [
@@ -137,7 +137,7 @@ lambda-web = { workspace = true, optional = true }
 log.workspace = true
 martin-core.workspace = true
 martin-tile-utils.workspace = true
-mbtiles = { workspace = true, optional = true }
+mbtiles.workspace = true
 moka.workspace = true
 num_cpus.workspace = true
 pmtiles = { workspace = true, optional = true }
