+++ martin/src/lib.rs
@@ -11,8 +11,6 @@ pub use utils::{IdResolver, MartinError, MartinResult, NO_MAIN_CACHE};
 
 #[cfg(feature = "cog")]
 pub mod cog;
-#[cfg(feature = "mbtiles")]
-pub mod mbtiles;
 #[cfg(feature = "postgres")]
 pub mod pg;
 #[cfg(feature = "pmtiles")]
